Restructure project with single run entrypoint
