+++ README.md
@@ -37,12 +37,12 @@ It offers a compact set of tools to keep groups clean while remaining easy to co
    - `SUPPORT_CHAT_URL`, `DEVELOPER_URL`, `PANEL_IMAGE_URL`
 3. Run the bot locally for testing
    ```bash
-   python3 oxygenbot.py
+   python3 run.py
    ```
    Keep it running using `screen`, `tmux` or a `systemd` service.
 
 4. To deploy on a server provide the same environment variables and execute
-   `sh start.sh`. The script runs `python oxygenbot.py` with logging enabled.
+   `sh start.sh`. The script runs `python run.py` with logging enabled.
 
 ## Render Deployment
 Create a new **Background Worker** on [Render](https://render.com) and use `render.yaml` for automatic configuration.
@@ -64,7 +64,7 @@ running in the background.
 4. Start the bot inside a persistent shell so it keeps running:
    ```bash
    screen -S oxygen
-   python3 oxygenbot.py
+   python3 run.py
    ```
    Detach with `Ctrl+A` then `D`. Reattach with `screen -r oxygen`.
