--- mybot/__init__.py
@@ -1,1 +1,0 @@
-__all__ = ['handlers', 'utils']
