--- mybot/requirements.txt
@@ -1,5 +1,0 @@
-pyrogram>=2.0.106
-motor>=3.4
-python-dotenv>=1.0.1
-tgcrypto>=1.2.5
-flask>=3.0
