--- mybot/utils/__init__.py
@@ -1,3 +1,0 @@
-from . import db, errors, perms, webhook, messages
-
-__all__ = ["db", "errors", "perms", "webhook", "messages"]
