--- oxeign/swagger/bots/bots_Dockerfile
@@ -1,12 +1,0 @@
-FROM python:3.12-slim
-ENV PYTHONUNBUFFERED=1
-WORKDIR /app
-COPY requirements.txt .
-# Install build tools required for packages like tgcrypto then remove them
-RUN apt-get update \ 
-    && apt-get install -y --no-install-recommends build-essential \ 
-    && pip install --no-cache-dir -r requirements.txt \ 
-    && apt-get purge -y --auto-remove build-essential \ 
-    && rm -rf /var/lib/apt/lists/*
-COPY . .
-CMD ["sh", "start.sh"]
