--- oxeign/swagger/bots/bots_handlers/bots_commands.py
@@ -1,14 +1,0 @@
-"""Shared list of bot admin commands."""
-
-COMMANDS = [
-    ("✅ /approve", "Approve a user"),
-    ("❌ /unapprove", "Revoke approval"),
-    ("📋 /viewapproved", "List approved users"),
-    ("🕒 /setautodelete <seconds>", "Enable auto delete"),
-    ("🤐 /mute", "Mute user"),
-    ("🚫 /kick", "Kick user"),
-    ("🔨 /ban", "Ban user"),
-    ("🌐 /biolink on | off", "Filter bio links"),
-    ("🔗 /linkfilter on | off", "Filter any link"),
-    ("✏️ /editmode on | off", "Enable/Disable edit filter"),
-]
