--- oxeign/swagger/harami/harami_env.example
@@ -1,9 +1,0 @@
-API_ID=123456
-API_HASH=your_api_hash
-BOT_TOKEN=123456:ABCDEF
-# Database
-MONGO_URI=
-MONGO_DB=oxygen
-# Optional
-OWNER_ID=1888832817
-BOT_USERNAME=
