--- oxeign/swagger/harami/harami_render.yaml
@@ -1,11 +1,0 @@
-services:
-  - type: web
-    name: guard-web
-    env: python
-    buildCommand: pip install -r requirements.txt
-    startCommand: python web.py
-  - type: worker
-    name: guard-bot
-    env: python
-    buildCommand: pip install -r requirements.txt
-    startCommand: python main.py
