--- oxeign/swagger/harami/harami_requirements.txt
@@ -1,4 +1,0 @@
-pyrogram>=2.0.106
-motor>=3.4
-python-dotenv>=1.0.1
-tgcrypto>=1.2.5
