--- oxeign/swagger/harami/harami_runtime.txt
@@ -1,1 +1,0 @@
-python-3.12.10
