--- oxeign/swagger/harami/harami_start.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-set -e
-exec python -u main.py
