--- oxeign/swagger/swagger/swagger_init.py
@@ -1,4 +1,0 @@
-from .handlers import register_all
-from .utils import db, errors, perms, webhook
-
-__all__ = ["register_all", "db", "errors", "perms", "webhook"]
