--- oxeign/swagger/yard/yard_utils/yard_utils_init.py
@@ -1,6 +1,0 @@
-# utils/__init__.py
-# Exposes key utility modules for use elsewhere in the project
-
-from . import db, errors, perms, webhook, messages
-
-__all__ = ["db", "errors", "perms", "webhook", "messages"]
