+++ run.py
@@ -1,4 +1,3 @@
-import asyncio
 import logging
 from pyrogram import Client, idle
 from pyrogram.enums import ParseMode
@@ -8,14 +7,12 @@ from handlers import register_all
 from utils.db import init_db, close_db
 from utils.webhook import delete_webhook
 
-# Configure logging
 logging.basicConfig(
     format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
-    level=getattr(logging, LOG_LEVEL.upper(), logging.INFO),
+    level=getattr(logging, LOG_LEVEL, logging.INFO),
 )
 logger = logging.getLogger(__name__)
 
-# Initialize the bot client
 bot = Client(
     "oxygen_bot",
     api_id=API_ID,
@@ -26,28 +23,17 @@ bot = Client(
 
 
 async def main() -> None:
-    logger.info("🚀 Starting OxygenBot...")
-
-    # Initialize MongoDB
+    logger.info("\uD83D\uDE80 Starting OxygenBot...")
     await init_db(MONGO_URI, MONGO_DB)
-    logger.info("✅ MongoDB connected.")
-
-    # Clear old webhooks
+    logger.info("\u2705 MongoDB connected.")
     await delete_webhook(BOT_TOKEN)
-    logger.info("🔌 Webhook deleted (if any). Switching to polling mode.")
-
-    # Register handlers from all modules
+    logger.info("\uD83D\uDD0C Webhook deleted (if any). Using polling mode.")
     register_all(bot)
-
-    # Start the bot and listen
-    async with bot:
-        logger.info("🤖 OxygenBot is live and listening...")
-        await idle()
-
-    # Graceful shutdown
+    logger.info("\uD83E\uDD16 Bot is live and listening...")
+    await idle()
     await close_db()
-    logger.info("🛑 Bot stopped cleanly.")
+    logger.info("\uD83D\uDD1A Bot stopped cleanly.")
 
 
 if __name__ == "__main__":
-    asyncio.run(main())
+    bot.run(main())
